Remove the old untyped Dynamo scanners
