+++ reindexer/reindex_worker/src/main/scala/uk/ac/wellcome/platform/reindex/reindex_worker/dynamo/ScanSpecScanner.scala
@@ -3,34 +3,34 @@ package uk.ac.wellcome.platform.reindex.reindex_worker.dynamo
 import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import com.amazonaws.services.dynamodbv2.document._
 import com.amazonaws.services.dynamodbv2.document.spec.ScanSpec
+import org.scanamo.DynamoFormat
 
 import scala.collection.JavaConverters._
 import scala.concurrent.{ExecutionContext, Future}
 
 /** Implements a wrapper for DynamoDB Scan operations using a ScanSpec.
-  *
-  * This wrapper provides a list of JSON strings, which can be sent directly
-  * to a downstream application.
-  *
-  * For the options allowed by ScanSpec, see:
-  * https://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/dynamodbv2/document/spec/ScanSpec.html
-  */
-class ScanSpecScanner(dynamoDBClient: AmazonDynamoDB)(
-  implicit ec: ExecutionContext) {
+ *
+ * This wrapper provides a list of JSON strings, which can be sent directly
+ * to a downstream application.
+ *
+ * For the options allowed by ScanSpec, see:
+ * https://docs.aws.amazon.com/AWSJavaSDK/latest/javadoc/com/amazonaws/services/dynamodbv2/document/spec/ScanSpec.html
+ *
+ * This is based on https://alexwlchan.net/2018/08/parallel-scan-scanamo/, although
+ * the APIs have changed slightly since that was written.
+ */
+trait ScanSpecScanner extends ItemParser {
+  implicit val dynamoClient: AmazonDynamoDB
+  implicit val ec: ExecutionContext
 
-  val dynamoDB = new DynamoDB(dynamoDBClient)
+  private val documentApiClient = new DynamoDB(dynamoClient)
 
-  /** Run a Scan specified by a ScanSpec.
-    *
-    * Note that this returns a Future[List], so results will be cached in-memory.
-    * Design your spec accordingly.
-    */
-  def scan(scanSpec: ScanSpec, tableName: String): Future[List[String]] = {
+  protected def scan[T](spec: ScanSpec)(tableName: String)(implicit format: DynamoFormat[T]): Future[Seq[T]] = {
+    val table = documentApiClient.getTable(tableName)
+
     for {
-      table <- Future.successful { dynamoDB.getTable(tableName) }
-      scanResult: ItemCollection[ScanOutcome] <- Future { table.scan(scanSpec) }
-      items: List[Item] = scanResult.asScala.toList
-      jsonStrings = items.map { _.toJSON }
-    } yield jsonStrings
+      items: Seq[Item] <- Future { table.scan(spec).asScala.toSeq }
+      result <- parseItems[T](items)
+    } yield result
   }
 }
