+++ reindexer/reindex_worker/src/main/scala/uk/ac/wellcome/platform/reindex/reindex_worker/services/RecordReader.scala
@@ -4,26 +4,22 @@ import com.amazonaws.services.dynamodbv2.AmazonDynamoDB
 import grizzled.slf4j.Logging
 import org.scanamo.DynamoFormat
 import uk.ac.wellcome.platform.reindex.reindex_worker.dynamo.{
-  BatchItemGetter,
-  MaxRecordsScanner,
   MultiItemGetter,
-  NewMaxRecordsScanner,
-  NewParallelScanner,
+  MaxRecordsScanner,
   ParallelScanner
 }
 import uk.ac.wellcome.platform.reindex.reindex_worker.models._
-import uk.ac.wellcome.storage.dynamo.DynamoConfig
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class NewRecordReader(
+class RecordReader(
   implicit
   dynamoClient: AmazonDynamoDB,
   ec: ExecutionContext
 ) extends Logging {
 
-  private val maxRecordsScanner = new NewMaxRecordsScanner()
-  private val parallelScanner = new NewParallelScanner()
+  private val maxRecordsScanner = new MaxRecordsScanner()
+  private val parallelScanner = new ParallelScanner()
   private val multiItemGetter = new MultiItemGetter()
 
   def findRecords[T](
@@ -41,33 +37,6 @@ class NewRecordReader(
 
       case SpecificReindexParameters(ids) =>
         multiItemGetter.get(ids)(tableName)
-    }
-  }
-}
-
-class RecordReader(
-  maxRecordsScanner: MaxRecordsScanner,
-  parallelScanner: ParallelScanner,
-  specificItemsGetter: BatchItemGetter
-) extends Logging {
-
-  def findRecordsForReindexing(
-    reindexParameters: ReindexParameters,
-    dynamoConfig: DynamoConfig): Future[List[String]] = {
-    debug(s"Finding records that need reindexing for $reindexParameters")
-
-    reindexParameters match {
-      case CompleteReindexParameters(segment, totalSegments) =>
-        parallelScanner
-          .scan(
-            segment = segment,
-            totalSegments = totalSegments
-          )(tableName = dynamoConfig.tableName)
-      case PartialReindexParameters(maxRecords) =>
-        maxRecordsScanner.scan(maxRecords = maxRecords)(
-          tableName = dynamoConfig.tableName)
-      case SpecificReindexParameters(ids) =>
-        specificItemsGetter.get(ids)(tableName = dynamoConfig.tableName)
     }
   }
 }
