--- reindexer/reindex_worker/src/test/scala/uk/ac/wellcome/platform/reindex/reindex_worker/fixtures/RecordReaderFixture.scala
@@ -1,12 +1,0 @@
-package uk.ac.wellcome.platform.reindex.reindex_worker.fixtures
-
-import uk.ac.wellcome.platform.reindex.reindex_worker.services.RecordReader
-
-trait RecordReaderFixture extends ReindexDynamoFixtures {
-  def createRecordReader: RecordReader =
-    new RecordReader(
-      maxRecordsScanner = createMaxRecordsScanner,
-      parallelScanner = createParallelScanner,
-      specificItemsGetter = createBatchItemGetter
-    )
-}
